requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
pyobjc-core>=10.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0

# Optional: Parakeet-MLX for Apple Silicon transcription
# Install with: pip install parakeet-mlx
parakeet-mlx>=0.5.0
